[Refactor] #348 - 공통 Enum 패키지 생성 및 구조 통합
- com.example.nexus.common.enums 패키지로 모든 Enum 이동

- 중복된 인벤토리 상태 Enum을 InventoryStatus로 통합

- 중복된 알림 타입 Enum을 NotificationType으로 통합

- 엔티티 모델의 Enum 참조 구조 및 import 업데이트
