+++ backend/src/main/java/com/example/nexus/common/enums/InventoryStatus.java
@@ -1,0 +1,7 @@
+package com.example.nexus.common.enums;
+
+public enum InventoryStatus {
+    NORMAL,
+    CRITICAL,
+    LOW
+}
