+++ backend/src/main/java/com/example/nexus/common/enums/NewsCollectTarget.java
@@ -1,4 +1,4 @@
-package com.example.nexus.newscollect.model;
+package com.example.nexus.common.enums;
 
 public enum NewsCollectTarget {
     HQ,
