--- backend/src/main/java/com/example/nexus/headinventory/model/HeadInventoryStatus.java
@@ -1,7 +1,0 @@
-package com.example.nexus.headinventory.model;
-
-public enum HeadInventoryStatus {
-    NORMAL,
-    CRITICAL,
-    LOW
-}
