--- backend/src/main/java/com/example/nexus/headnotification/model/Status.java
@@ -1,7 +1,0 @@
-package com.example.nexus.headnotification.model;
-
-public enum Status {
-    NORMAL,
-    CRITICAL,
-    LOW
-}
