--- backend/src/main/java/com/example/nexus/headnotification/model/Type.java
@@ -1,8 +1,0 @@
-package com.example.nexus.headnotification.model;
-
-public enum Type {
-    LOW,
-    CRITICAL,
-    ABNORMAL,
-    DELAY
-}
