--- backend/src/main/java/com/example/nexus/posstoreinventory/model/PosStoreInventoryStatus.java
@@ -1,7 +1,0 @@
-package com.example.nexus.posstoreinventory.model;
-
-public enum PosStoreInventoryStatus {
-    NORMAL,
-    CRITICAL,
-    LOW
-}
