+++ backend/src/main/java/com/example/nexus/storeinventory/model/StoreInventory.java
@@ -1,5 +1,6 @@
 package com.example.nexus.storeinventory.model;
 
+import com.example.nexus.common.enums.InventoryStatus;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
@@ -33,7 +34,7 @@ public class StoreInventory {
 
     @Enumerated(EnumType.STRING)
     @Column(name = "status", nullable = false)
-    private StoreInventoryStatus status = StoreInventoryStatus.NORMAL;
+    private InventoryStatus status = InventoryStatus.NORMAL;
 
     @Column(name = "manufactured_date", nullable = false)
     private LocalDateTime manufacturedDate;
