--- backend/src/main/java/com/example/nexus/storeinventory/model/StoreInventoryStatus.java
@@ -1,7 +1,0 @@
-package com.example.nexus.storeinventory.model;
-
-public enum StoreInventoryStatus {
-    NORMAL,
-    CRITICAL,
-    LOW
-}
